refactor: reorganizes logic files
