+++ src/logic/downloadBlob.js
@@ -1,10 +1,6 @@
-import RSAKey from 'react-native-rsa'
-
-const rsa = new RSAKey()
-
 import RNFetchBlob from 'rn-fetch-blob'
 
-const downloadBlob = (url, type = 'jpg') =>
+export const downloadBlob = (url, type = 'jpg') =>
   new Promise((resolve, reject) => {
     RNFetchBlob.config({
       fileCache: true,
@@ -17,5 +13,3 @@ const downloadBlob = (url, type = 'jpg') =>
         reject(err)
       })
   })
-
-export {rsa, downloadBlob}
