+++ src/logic/rsa.js
@@ -1,0 +1,3 @@
+import RSAKey from 'react-native-rsa'
+
+export const rsa = new RSAKey()
